+++ CotEditor/Sources/StepperNumberField.swift
@@ -28,38 +28,70 @@ import SwiftUI
 struct StepperNumberField: View {
     
     @Binding private var value: Int
+    private var defaultValue: Int?
     private var bounds: ClosedRange<Int>
     private var step: Int.Stride
-    private var prompt: Text?
+    private var promptText: LocalizedStringKey?
+    private var fieldWidth: CGFloat? = 32
     
     
     /// Create a text field editing numerical value with a stepper.
     ///
     /// - Parameters:
     ///   - value: The underlying value to edit.
+    ///   - defaultValue: The default value when the field is empty.
     ///   - bounds: A closed range that describes the upper and lower bounds permitted by the stepper.
     ///   - step: The amount to increment or decrement value each time the user clicks or taps the stepper.
     ///   - prompt: A Text which provides users with guidance on what to type into the text field.
-    init(value: Binding<Int>, in bounds: ClosedRange<Int>, step: Int.Stride = 1, prompt: Text? = nil) {
+    init(value: Binding<Int>, default defaultValue: Int? = nil, in bounds: ClosedRange<Int>, step: Int.Stride = 1, prompt: LocalizedStringKey? = nil) {
         
         self._value = value
+        self.defaultValue = defaultValue
         self.bounds = bounds
         self.step = step
-        self.prompt = prompt
+        self.promptText = prompt
     }
     
     
     var body: some View {
         
         HStack(spacing: 4) {
-            TextField("", value: $value, format: .ranged(self.bounds), prompt: self.prompt)
-                .monospacedDigit()
-                .multilineTextAlignment(.trailing)
-                .frame(width: 32)
-            Stepper("", value: $value, in: self.bounds, step: self.step)
+            TextField(value: $value, format: .ranged(self.bounds, defaultValue: self.defaultValue), prompt: self.prompt) {
+                EmptyView()
+            }
+            .monospacedDigit()
+            .multilineTextAlignment(.trailing)
+            .frame(width: self.fieldWidth)
+            
+            Stepper(value: $value, in: self.bounds, step: self.step) {
+                EmptyView()
+            }
         }
         .labelsHidden()
         .fixedSize()
+    }
+    
+    
+    private var prompt: Text? {
+        
+        if let defaultValue {
+            Text(defaultValue, format: .number)
+        } else if let promptText {
+            Text(promptText)
+        } else {
+             nil
+        }
+    }
+    
+    
+    /// Set the input field width to the specified size.
+    ///
+    /// - Parameter fieldWidth: The field width.
+    func fieldWidth(_ fieldWidth: CGFloat?) -> some View {
+        
+        var view = self
+        view.fieldWidth = fieldWidth
+        return view
     }
 }